wip: ants distribution cleanup
